Added date to media information.
Media.xml contains a tag with the air date of the program. This is
now also extracted and stored in table "media". This will allow
better control when deciding what to download in an abo.
Note however, that a little percentage (181 out of 40K+ as of today)
of programs don't have the tag <Datum> defined.